Parser

Documentation

data JValue

Constructors

JNull

JBool Bool

JNumber Double

JString String

JArray [JValue]

JObject [(String, JValue)]

type Parser a = Parsec String () a

Type alias for Parsec that takes a String as input and outputs a Does not utilize the user state, as given by the ()

jNullParser :: Parser JValue

Only parses the keyword null with no whitespace

>>> parseTest jNullParser "null"
JNull

>>> parseTest jNullParser "foo"
parse error at (line 1, column 1):
unexpected "f"
expecting "null"

jBoolParser :: Parser JValue

Parses either true or false with no whitespace

>>> parseTest jBoolParser "true"
JBool True

>>> parseTest jBoolParser "bar"
parse error at (line 1, column 1):
unexpected "b"
expecting "true" or "false"

jNumberParser :: Parser JValue

Parses a double or integer value

>>> parseTest jNumberParser "-23.6"
JNumber (-23.6)

>>> parseTest jNumberParser "42"
JNumber 42.0

>>> parseTest jNumberParser "baz"
parse error at (line 1, column 1):
unexpected "b"
expecting "-", float or natural

jStringParser :: Parser JValue

Parses a string with double quotes around it, returning what is inside

>>> parseTest jStringParser  "\"hello\""
JString "hello"

This example is missing the double quotes: >>> parseTest jStringParser "hello" parse error at (line 1, column 1): unexpected "h" expecting literal string

stringLiteral :: Parser String

Parses any string with double quotes around it, returning what is inside Currently only works for whitespace escaped characters such as 'n' Other escaped characters will remain in their raw form

jArrayParser :: Parser JValue

Parses an array surrounded by brackets and separated by commas and whitespace

>>> parseTest jArrayParser  "[1, 2, 3, 4]"
JArray [JNumber 1.0,JNumber 2.0,JNumber 3.0,JNumber 4.0]

Since it is a JSON, there can be different types within the array: >>> parseTest jArrayParser "[1, true, null, "hello"]" JArray [JNumber 1.0,JBool True,JNull,JString "hello"]

>>> parseTest jArrayParser  "1, true, null, \"hello\"]"
parse error at (line 1, column 1):
unexpected "1"
expecting "["

jObjectParser :: Parser JValue

Parses an object of String and JValue pairs surrounded by braces and separated by commas and whitespace

>>> parseTest jObjectParser "{ \"first\" : 24, \"second\" : null }"
JObject [("first",JNumber 24.0),("second",JNull)]

>>> parseTest jObjectParser " \"first\" : 24, \"second\" : null }"
parse error at (line 1, column 1):
unexpected " "
expecting "{"

jValueParser :: Parser JValue

Combines all parsers into a single JSON parser that allows for any amount of whitespace between them

runJsonParser :: String -> Parser JValue -> Either ParseError JValue

Runs the given parser on a string and returns the resulting JValue or error

parseJson :: String -> Either ParseError JValue

Runs jValueParser on the given string and returns the resulting JValue or an error